Complete Advent of code 2021 day 1 task 2
The task was to compare sliding windows of values, and if the sum of the
values in the window was larger than the previous window, then increase
the count.

Spotted that the only numbers you actually need to compare are the
current one and the oldest one in the window. Used a VecDeque as a queue
to store the items from the previous window and we can then pop the
oldest value, check if it is larger or smaller than the current value,
then push the current value into the queue to prepare for the next
iteration.

Could we replace the VecDeque with a fixed size queue? We know the size
of the sliding window for this task, and we don't need to ever grow this
queue larger than that.

I don't like making the VecDeque contain Options. It feels a little
untidy, but I wanted some way to spot I was on the first 3 iterations
and not count these in the result.