Use str.parse() instead of calling from_str directly
This is really neat. The type inference works out what type to parse to
for me automatically. I assume it's tracking this all the way from the
SubState type that I am summing to. The compiler is clever enough to
work out that the only type that sums to a SubType is the Movement enum
and parses the string to that without me having to explicitly state the
type.