Advent of Code day 8 complete
Not the neatest code ever, but it works.
A bunch of set-based shenanigans. Some nastiness because I can't use a
HashSet as the key to a HashMap. Would be more convenient if I could do
a direct look up with the set of characters rather than having to loop
over all the mappings in the cipher.

I do like that Rust supports infix operators for Set operations out of
the box.